MHR staff death transfer validation updates.
