+++ mhr-api/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "mhr-api"
-version = "2.0.4"
+version = "2.0.5"
 description = ""
 authors = ["dlovett <doug@daxiom.com>"]
 license = "BSD 3"
